fix: Update metrics increment type for duplicate evidence count
- Changed the increment type for the `api_evidence_duplicates` metric from `u64` to `f64` to ensure compatibility with potential floating-point operations in future metric calculations.
- This adjustment enhances the accuracy and flexibility of the metrics tracking for evidence deduplication.

+++ recap-worker/recap-worker/src/api/fetch.rs
@@ -168,7 +168,7 @@ pub(crate) async fn get_7days_recap(State(state): State<AppState>) -> impl IntoR
             if removed > 0 {
                 metrics
                     .api_evidence_duplicates
-                    .inc_by(removed as u64);
+                    .inc_by(removed as f64);
             }
             deduped
         };
